Create Group(), Groups() & Groupdict()
Problem:
The task is to find the first consecutive alphanumeric character in a string (S) (i.e., the character appears twice in a row). If no such character is found, -1 should be printed.
For example: if the input is "hello", the output is (l) (because (l) is repeated).
In this code, a for loop with some conditions is used to search for the first consecutive character in the text entered by the user. The program begins by reading a text string from the user using input(). Next, a for loop is used to iterate over each character in the string (except the last) using range(len(string) - 1), where each character is compared to the next using the condition string[i] == string[i + 1]. The condition string[i]. isalnum() is also used to verify that the repeated character is a letter or a number (i.e., an alphanumeric symbol). If both conditions are true, the character is printed, the variable found is set to True, and the loop is stopped using break. If no duplicate characters are found, -1 is printed after the loop ends, and the if not found condition is used for that.