Create Input()
Problem:
The goal is to check the value of a polynomial when a specific value is substituted for the variable x, and compare the result with the given value of k.
Inputs:
The value of x and the value of k (integers).
Polynomial P (as a string, such as "x**2 + 3*x + 2").
Required:
Calculate P(x) (i.e., substitute x into the equation).
If the result is equal to k, print True.
Otherwise, print False.

In this code, the input() function is used to get user input in two steps. First, the user enters two numbers (x and k) separated by a space. The code splits the input using .split() and converts both values to integers using int(). Then, the user enters a math expression (like "x**2 + 3*x + 1") that includes the variable x.

The program replaces the letter x in the expression with the actual number using .replace(). After that, it uses eval() to calculate the result of the expression. Finally, it checks if the result is equal to k. If it is, it prints True; otherwise, it prints False.