Create Piling Up!
Problem:
You have a set of cubes arranged horizontally. You must form a vertical stack from them if:
When placing a cube on top of another, the size of the top cube must be smaller than or equal to the bottom cube.
At each step, you can only take either the leftmost or rightmost cube from the current horizontal arrangement.
You need to check whether a correct vertical stack can be formed using these rules.
Example:
[1, 2, 3, 8, 7] → No
(A correct stack cannot be formed because 8 and 7 are greater than 1 after taking 7 first.)

This code uses a simplified set of Python tools. The input() function is used to read data from the user, and the int() function is used to convert the input values to integers. Map() and list() are used to convert a string of numbers to a list of numeric values. The code uses a for loop to repeat the process according to the number of states, and a while loop to process the cubes within each state as long as there are cubes left. We also used if, elif, and else conditions to compare the cubes from both ends and determine which one can be placed in the heap according to the rule. To remove cubes from the list, we used pop(0) to remove the first element and pop() to remove the last element. float('inf') is used to represent the beginning of the heap with a very large value so that we can compare the cubes to it. Finally, we used print() to display the result: "Yes" if the order is possible, or "No" if it is not.